Flash Card
This is very useful if you want to be good at a particular foreign language! Every language has common words So here I have taken all this make it in the form of CSV! In this first german language will show and after 3 seconds it will flip card and show its meaning! Also, there are 2 buttons if you know the meaning of that word so you can click on the right button within 3 seconds after that word will not show to you again and if you clicked the cross button it will remain in a new CSV file and after a period of time, it will come again to make your dictionary strong!